requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.14.2
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
